refactor(service): 用户静音/耳聋等发送收回 core/net 的 sender 方法，并修掉最后一处编译告警

+++ app/src/test/java/dev/woms/mumdroid/UdpProtocolTest.kt
@@ -254,7 +254,7 @@ class UdpProtocolTest {
         assertEquals(1, packet[0].toInt()) // Ping header byte
         val ts = dev.woms.mumdroid.core.net.ProtoUdpCodec.decodePing(packet.copyOfRange(1, packet.size))
         assertNotNull(ts)
-        assertEquals(987654321L, ts!!.toLong())
+        assertEquals(987654321L, ts!!)
     }
 
     @Test
